chore: extracted some code out of the card type
